examples/debug_tutorial.go
package examples

import (
	"context"
	"fmt"
	"math/big"
)

// SimpleArbitrage represents a basic arbitrage opportunity
type SimpleArbitrage struct {
	TokenA      string
	TokenB      string
	AmountIn    *big.Int
	AmountOut   *big.Int
	Profit      *big.Int
	Path        []string
	GasEstimate uint64
}

// CalculateProfit demonstrates a simple profit calculation with multiple steps
func CalculateProfit(ctx context.Context, arb *SimpleArbitrage) error {
	// Step 1: Validate input
	if arb.AmountIn == nil || arb.AmountIn.Sign() <= 0 {
		return fmt.Errorf("invalid input amount")
	}

	// Step 2: Calculate initial exchange
	exchangeRate := big.NewInt(105)  // 1.05 rate
	intermediateAmount := new(big.Int).Mul(arb.AmountIn, exchangeRate)
	intermediateAmount.Div(intermediateAmount, big.NewInt(100))

	// Step 3: Calculate second exchange
	finalRate := big.NewInt(103)  // 1.03 rate
	finalAmount := new(big.Int).Mul(intermediateAmount, finalRate)
	finalAmount.Div(finalAmount, big.NewInt(100))

	// Step 4: Calculate profit
	arb.AmountOut = finalAmount
	arb.Profit = new(big.Int).Sub(finalAmount, arb.AmountIn)

	// Step 5: Estimate gas (simplified)
	arb.GasEstimate = 150000  // Example gas estimate

	// Step 6: Update path
	arb.Path = []string{
		fmt.Sprintf("Start with %v TokenA", arb.AmountIn),
		fmt.Sprintf("Exchange to %v TokenB", intermediateAmount),
		fmt.Sprintf("Exchange back to %v TokenA", finalAmount),
	}

	return nil
}

// ExecuteArbitrage simulates executing the arbitrage
func ExecuteArbitrage(ctx context.Context, arb *SimpleArbitrage) error {
	// Step 1: Check profitability
	minProfit := big.NewInt(1000000)  // Minimum profit threshold
	if arb.Profit.Cmp(minProfit) < 0 {
		return fmt.Errorf("profit %v below minimum threshold %v", arb.Profit, minProfit)
	}

	// Step 2: Simulate transaction execution
	for i, step := range arb.Path {
		select {
		case <-ctx.Done():
			return ctx.Err()
		default:
			fmt.Printf("Step %d: %s\n", i+1, step)
		}
	}

	return nil
}
